# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AssessmentAnswersRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          }
        },
        "required": [
          "answers"
        ],
        "title": "AssessmentAnswersRequest",
        "type": "object"
      },
      "AssessmentStartRequest": {
        "properties": {
          "form": {
            "enum": [
              "A",
              "B"
            ],
            "title": "Form",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "mode": {
            "default": "test",
            "enum": [
              "test",
              "practice"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "form"
        ],
        "title": "AssessmentStartRequest",
        "type": "object"
      },
      "AssessmentSubmitRequest": {
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          }
        },
        "title": "AssessmentSubmitRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend for the Quantum Learning Laboratory.\n\nThe quantum engine computes, the authored evaluator grades, and the tutor\nteaches. Private rubrics and answer keys are never returned by any route in\nthis contract.",
    "title": "Quantum Learning Laboratory API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assessments/attempts/{attempt_id}": {
      "patch": {
        "description": "Answers are saved server-side so a reload or a crash does not lose them.",
        "operationId": "save_answers_api_v1_assessments_attempts__attempt_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentAnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Answers Api V1 Assessments Attempts  Attempt Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Answers",
        "tags": [
          "assessment"
        ]
      }
    },
    "/api/v1/assessments/attempts/{attempt_id}/convert-to-practice": {
      "post": {
        "description": "The platform performs this transition and records the assistance, not the tutor.",
        "operationId": "convert_to_practice_api_v1_assessments_attempts__attempt_id__convert_to_practice_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Convert To Practice Api V1 Assessments Attempts  Attempt Id  Convert To Practice Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Convert To Practice",
        "tags": [
          "assessment"
        ]
      }
    },
    "/api/v1/assessments/attempts/{attempt_id}/submit": {
      "post": {
        "operationId": "submit_attempt_api_v1_assessments_attempts__attempt_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "title": "Attempt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Attempt Api V1 Assessments Attempts  Attempt Id  Submit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Attempt",
        "tags": [
          "assessment"
        ]
      }
    },
...
    "/api/v1/assessments/chapter-1/attempts": {
      "post": {
        "operationId": "start_attempt_api_v1_assessments_chapter_1_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Attempt Api V1 Assessments Chapter 1 Attempts Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Attempt",
        "tags": [
          "assessment"
        ]
      }
    },
...
  }
}